requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
faster-whisper==1.0.3
ffmpeg-python==0.2.0
supabase==2.6.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2

# Optional (install manually if needed)
# transformers==4.44.2
pydub==0.25.1
ctranslate2==4.6.0

# For CPU-only torch (if needed on Render)
# If Render build fails with torch, consider using:
# torch==2.3.1+cpu ; platform_system != 'Darwin'
# --extra-index-url https://download.pytorch.org/whl/cpu
